Consolidate FIO scenes into a folder
