+++ src/__tests__/scenes/FioAddressRegisterScene.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect, it } from '@jest/globals'
 import * as React from 'react'
 import { createRenderer } from 'react-test-renderer/shallow'
 
-import { FioAddressRegister } from '../../components/scenes/FioAddressRegisterScene'
+import { FioAddressRegister } from '../../components/scenes/Fio/FioAddressRegisterScene'
 import { getTheme } from '../../components/services/ThemeContext'
 import { fakeNavigation } from '../../util/fake/fakeNavigation'
