+++ src/__tests__/scenes/FioAddressRegisteredScene.test.tsx
@@ -2,7 +2,7 @@ import { describe, expect, it } from '@jest/globals'
 import * as React from 'react'
 import { createRenderer } from 'react-test-renderer/shallow'
 
-import { FioAddressRegistered } from '../../components/scenes/FioAddressRegisteredScene'
+import { FioAddressRegistered } from '../../components/scenes/Fio/FioAddressRegisteredScene'
 import { getTheme } from '../../components/services/ThemeContext'
 import { fakeNavigation } from '../../util/fake/fakeNavigation'
 import { fakeNonce } from '../../util/fake/fakeNonce'
