+++ src/components/Main.tsx
@@ -50,22 +50,22 @@ import { DefaultFiatSettingScene as DefaultFiatSettingSceneComponent } from './s
 import { EdgeLoginScene as EdgeLoginSceneComponent } from './scenes/EdgeLoginScene'
 import { EditTokenScene as EditTokenSceneComponent } from './scenes/EditTokenScene'
 import { ExtraTabScene as ExtraTabSceneComponent } from './scenes/ExtraTabScene'
-import { FioAddressDetailsScene as FioAddressDetailsSceneComponent } from './scenes/FioAddressDetailsScene'
-import { FioAddressListScene as FioAddressListSceneComponent } from './scenes/FioAddressListScene'
-import { FioAddressRegisteredScene as FioAddressRegisteredSceneComponent } from './scenes/FioAddressRegisteredScene'
-import { FioAddressRegisterScene as FioAddressRegisterSceneComponent } from './scenes/FioAddressRegisterScene'
-import { FioAddressRegisterSelectWalletScene as FioAddressRegisterSelectWalletSceneComponent } from './scenes/FioAddressRegisterSelectWalletScene'
-import { FioAddressSettingsScene as FioAddressSettingsSceneComponent } from './scenes/FioAddressSettingsScene'
-import { FioConnectWalletConfirmScene as FioConnectWalletConfirmSceneComponent } from './scenes/FioConnectWalletConfirmScene'
-import { FioDomainRegisterScene as FioDomainRegisterSceneComponent } from './scenes/FioDomainRegisterScene'
-import { FioDomainRegisterSelectWalletScene as FioDomainRegisterSelectWalletSceneComponent } from './scenes/FioDomainRegisterSelectWalletScene'
-import { FioDomainSettingsScene as FioDomainSettingsSceneComponent } from './scenes/FioDomainSettingsScene'
-import { FioNameConfirmScene as FioNameConfirmSceneComponent } from './scenes/FioNameConfirmScene'
-import { FioRequestConfirmationScene as FioRequestConfirmationSceneComponent } from './scenes/FioRequestConfirmationScene'
-import { FioRequestListScene as FioRequestListSceneComponent } from './scenes/FioRequestListScene'
-import { FioSentRequestDetailsScene as FioSentRequestDetailsSceneComponent } from './scenes/FioSentRequestDetailsScene'
-import { FioStakingChangeScene as FioStakingChangeSceneComponent } from './scenes/FioStakingChangeScene'
-import { FioStakingOverviewScene as FioStakingOverviewSceneComponent } from './scenes/FioStakingOverviewScene'
+import { FioAddressDetailsScene as FioAddressDetailsSceneComponent } from './scenes/Fio/FioAddressDetailsScene'
+import { FioAddressListScene as FioAddressListSceneComponent } from './scenes/Fio/FioAddressListScene'
+import { FioAddressRegisteredScene as FioAddressRegisteredSceneComponent } from './scenes/Fio/FioAddressRegisteredScene'
+import { FioAddressRegisterScene as FioAddressRegisterSceneComponent } from './scenes/Fio/FioAddressRegisterScene'
+import { FioAddressRegisterSelectWalletScene as FioAddressRegisterSelectWalletSceneComponent } from './scenes/Fio/FioAddressRegisterSelectWalletScene'
+import { FioAddressSettingsScene as FioAddressSettingsSceneComponent } from './scenes/Fio/FioAddressSettingsScene'
+import { FioConnectWalletConfirmScene as FioConnectWalletConfirmSceneComponent } from './scenes/Fio/FioConnectWalletConfirmScene'
+import { FioDomainRegisterScene as FioDomainRegisterSceneComponent } from './scenes/Fio/FioDomainRegisterScene'
+import { FioDomainRegisterSelectWalletScene as FioDomainRegisterSelectWalletSceneComponent } from './scenes/Fio/FioDomainRegisterSelectWalletScene'
+import { FioDomainSettingsScene as FioDomainSettingsSceneComponent } from './scenes/Fio/FioDomainSettingsScene'
+import { FioNameConfirmScene as FioNameConfirmSceneComponent } from './scenes/Fio/FioNameConfirmScene'
+import { FioRequestConfirmationScene as FioRequestConfirmationSceneComponent } from './scenes/Fio/FioRequestConfirmationScene'
+import { FioRequestListScene as FioRequestListSceneComponent } from './scenes/Fio/FioRequestListScene'
+import { FioSentRequestDetailsScene as FioSentRequestDetailsSceneComponent } from './scenes/Fio/FioSentRequestDetailsScene'
+import { FioStakingChangeScene as FioStakingChangeSceneComponent } from './scenes/Fio/FioStakingChangeScene'
+import { FioStakingOverviewScene as FioStakingOverviewSceneComponent } from './scenes/Fio/FioStakingOverviewScene'
 import { GettingStartedScene } from './scenes/GettingStartedScene'
 import { GuiPluginListScene as GuiPluginListSceneComponent } from './scenes/GuiPluginListScene'
 import { GuiPluginViewScene as GuiPluginViewSceneComponent } from './scenes/GuiPluginViewScene'
