+++ src/components/scenes/Fio/FioAddressDetailsScene.tsx
@@ -3,17 +3,17 @@ import * as React from 'react'
 import { Alert } from 'react-native'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 
-import s from '../../locales/strings'
-import { ConnectWalletsConnector as ConnectWallets } from '../../modules/FioAddress/components/ConnectWallets'
-import { BUNDLED_TXS_AMOUNT_ALERT, findWalletByFioAddress } from '../../modules/FioAddress/util'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { showError } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { EdgeText } from '../themed/EdgeText'
-import { SettingsHeaderRow } from '../themed/SettingsHeaderRow'
-import { SettingsTappableRow } from '../themed/SettingsTappableRow'
+import s from '../../../locales/strings'
+import { ConnectWalletsConnector as ConnectWallets } from '../../../modules/FioAddress/components/ConnectWallets'
+import { BUNDLED_TXS_AMOUNT_ALERT, findWalletByFioAddress } from '../../../modules/FioAddress/util'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { showError } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { EdgeText } from '../../themed/EdgeText'
+import { SettingsHeaderRow } from '../../themed/SettingsHeaderRow'
+import { SettingsTappableRow } from '../../themed/SettingsTappableRow'
 
 interface StateProps {
   fioWallets: EdgeCurrencyWallet[]
