+++ src/components/scenes/Fio/FioAddressListScene.tsx
@@ -3,22 +3,22 @@ import * as React from 'react'
 import { ActivityIndicator, Image, ScrollView, View } from 'react-native'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 
-import { refreshAllFioAddresses } from '../../actions/FioAddressActions'
-import fioAddressLogo from '../../assets/images/fio/fio_logo.png'
-import { Fontello } from '../../assets/vector'
-import s from '../../locales/strings'
-import { FioNameRow } from '../../modules/FioAddress/components/FioName'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp } from '../../types/routerTypes'
-import { FioAddress, FioDomain } from '../../types/types'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { FullScreenLoader } from '../progress-indicators/FullScreenLoader'
-import { showError } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { ClickableText } from '../themed/ClickableText'
-import { EdgeText } from '../themed/EdgeText'
-import { Fade } from '../themed/Fade'
-import { SceneHeader } from '../themed/SceneHeader'
+import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
+import fioAddressLogo from '../../../assets/images/fio/fio_logo.png'
+import { Fontello } from '../../../assets/vector'
+import s from '../../../locales/strings'
+import { FioNameRow } from '../../../modules/FioAddress/components/FioName'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp } from '../../../types/routerTypes'
+import { FioAddress, FioDomain } from '../../../types/types'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { FullScreenLoader } from '../../progress-indicators/FullScreenLoader'
+import { showError } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { ClickableText } from '../../themed/ClickableText'
+import { EdgeText } from '../../themed/EdgeText'
+import { Fade } from '../../themed/Fade'
+import { SceneHeader } from '../../themed/SceneHeader'
 
 interface LocalState {
   initLoading: boolean
