+++ src/components/scenes/Fio/FioAddressRegisterScene.tsx
@@ -3,28 +3,28 @@ import * as React from 'react'
 import { Image, LayoutChangeEvent, ScrollView, View } from 'react-native'
 import { sprintf } from 'sprintf-js'
 
-import { createFioWallet } from '../../actions/FioAddressActions'
-import { Fontello } from '../../assets/vector'
-import { FIO_ADDRESS_DELIMITER, FIO_DOMAIN_DEFAULT, FIO_STR } from '../../constants/WalletAndCurrencyConstants'
-import s from '../../locales/strings'
-import { DomainListModal } from '../../modules/FioAddress/components/DomainListModal'
-import { checkIsDomainPublic } from '../../modules/FioAddress/util'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp } from '../../types/routerTypes'
-import { FioDomain, FioPublicDomain } from '../../types/types'
-import { getWalletName } from '../../util/CurrencyWalletHelpers'
-import { openLink } from '../../util/utils'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { TextInputModal } from '../modals/TextInputModal'
-import { WalletListModal, WalletListResult } from '../modals/WalletListModal'
-import { Airship, showError, showToast } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { ClickableText } from '../themed/ClickableText'
-import { EdgeText } from '../themed/EdgeText'
-import { FormError } from '../themed/FormError'
-import { MainButton } from '../themed/MainButton'
-import { SceneHeader } from '../themed/SceneHeader'
-import { Tile } from '../tiles/Tile'
+import { createFioWallet } from '../../../actions/FioAddressActions'
+import { Fontello } from '../../../assets/vector'
+import { FIO_ADDRESS_DELIMITER, FIO_DOMAIN_DEFAULT, FIO_STR } from '../../../constants/WalletAndCurrencyConstants'
+import s from '../../../locales/strings'
+import { DomainListModal } from '../../../modules/FioAddress/components/DomainListModal'
+import { checkIsDomainPublic } from '../../../modules/FioAddress/util'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp } from '../../../types/routerTypes'
+import { FioDomain, FioPublicDomain } from '../../../types/types'
+import { getWalletName } from '../../../util/CurrencyWalletHelpers'
+import { openLink } from '../../../util/utils'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { TextInputModal } from '../../modals/TextInputModal'
+import { WalletListModal, WalletListResult } from '../../modals/WalletListModal'
+import { Airship, showError, showToast } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { ClickableText } from '../../themed/ClickableText'
+import { EdgeText } from '../../themed/EdgeText'
+import { FormError } from '../../themed/FormError'
+import { MainButton } from '../../themed/MainButton'
+import { SceneHeader } from '../../themed/SceneHeader'
+import { Tile } from '../../tiles/Tile'
 
 interface State {
   selectedWallet: EdgeCurrencyWallet | null
