+++ src/components/scenes/Fio/FioAddressRegisteredScene.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react'
 import { Image, View } from 'react-native'
 
-import { formatDate } from '../../locales/intl'
-import s from '../../locales/strings'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { FormattedText } from '../legacy/FormattedText/FormattedText.ui'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { MainButton } from '../themed/MainButton'
+import { formatDate } from '../../../locales/intl'
+import s from '../../../locales/strings'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { FormattedText } from '../../legacy/FormattedText/FormattedText.ui'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { MainButton } from '../../themed/MainButton'
 
 interface OwnProps {
   navigation: NavigationProp<'fioAddressRegisterSuccess'>
