+++ src/components/scenes/Fio/FioAddressSettingsScene.tsx
@@ -1,19 +1,19 @@
 import { EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
 
-import { refreshAllFioAddresses } from '../../actions/FioAddressActions'
-import s from '../../locales/strings'
-import { FioActionSubmit } from '../../modules/FioAddress/components/FioActionSubmit'
-import { addBundledTxs, getAddBundledTxsFee, getTransferFee } from '../../modules/FioAddress/util'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { ButtonsModal } from '../modals/ButtonsModal'
-import { Airship, showError, showToast } from '../services/AirshipInstance'
-import { ThemeProps, withTheme } from '../services/ThemeContext'
-import { EdgeText } from '../themed/EdgeText'
-import { MainButton } from '../themed/MainButton'
-import { Tile } from '../tiles/Tile'
+import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
+import s from '../../../locales/strings'
+import { FioActionSubmit } from '../../../modules/FioAddress/components/FioActionSubmit'
+import { addBundledTxs, getAddBundledTxsFee, getTransferFee } from '../../../modules/FioAddress/util'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { ButtonsModal } from '../../modals/ButtonsModal'
+import { Airship, showError, showToast } from '../../services/AirshipInstance'
+import { ThemeProps, withTheme } from '../../services/ThemeContext'
+import { EdgeText } from '../../themed/EdgeText'
+import { MainButton } from '../../themed/MainButton'
+import { Tile } from '../../tiles/Tile'
 
 interface LocalState {
   showAddBundledTxs: boolean
