+++ src/components/scenes/Fio/FioConnectWalletConfirmScene.tsx
@@ -1,20 +1,20 @@
 import * as React from 'react'
 import { ScrollView } from 'react-native'
 
-import s from '../../locales/strings'
-import { FIO_NO_BUNDLED_ERR_CODE, updatePubAddressesForFioAddress } from '../../modules/FioAddress/util'
-import { CcWalletMap } from '../../reducers/FioReducer'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { FioConnectionWalletItem } from '../../types/types'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { ButtonsModal } from '../modals/ButtonsModal'
-import { Airship, showError, showToast } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { EdgeText } from '../themed/EdgeText'
-import { Slider } from '../themed/Slider'
-import { Radio } from '../themed/ThemedButtons'
-import { Tile } from '../tiles/Tile'
+import s from '../../../locales/strings'
+import { FIO_NO_BUNDLED_ERR_CODE, updatePubAddressesForFioAddress } from '../../../modules/FioAddress/util'
+import { CcWalletMap } from '../../../reducers/FioReducer'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { FioConnectionWalletItem } from '../../../types/types'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { ButtonsModal } from '../../modals/ButtonsModal'
+import { Airship, showError, showToast } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { EdgeText } from '../../themed/EdgeText'
+import { Slider } from '../../themed/Slider'
+import { Radio } from '../../themed/ThemedButtons'
+import { Tile } from '../../tiles/Tile'
 
 interface State {
   acknowledge: boolean
