+++ src/components/scenes/Fio/FioDomainRegisterScene.tsx
@@ -3,22 +3,22 @@ import * as React from 'react'
 import { LayoutChangeEvent, ScrollView, View } from 'react-native'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 
-import { createFioWallet } from '../../actions/FioAddressActions'
-import { FIO_STR } from '../../constants/WalletAndCurrencyConstants'
-import s from '../../locales/strings'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp } from '../../types/routerTypes'
-import { getWalletName } from '../../util/CurrencyWalletHelpers'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { TextInputModal } from '../modals/TextInputModal'
-import { WalletListModal, WalletListResult } from '../modals/WalletListModal'
-import { Airship, showError, showToast } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { EdgeText } from '../themed/EdgeText'
-import { FormError } from '../themed/FormError'
-import { MainButton } from '../themed/MainButton'
-import { SceneHeader } from '../themed/SceneHeader'
-import { Tile } from '../tiles/Tile'
+import { createFioWallet } from '../../../actions/FioAddressActions'
+import { FIO_STR } from '../../../constants/WalletAndCurrencyConstants'
+import s from '../../../locales/strings'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp } from '../../../types/routerTypes'
+import { getWalletName } from '../../../util/CurrencyWalletHelpers'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { TextInputModal } from '../../modals/TextInputModal'
+import { WalletListModal, WalletListResult } from '../../modals/WalletListModal'
+import { Airship, showError, showToast } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { EdgeText } from '../../themed/EdgeText'
+import { FormError } from '../../themed/FormError'
+import { MainButton } from '../../themed/MainButton'
+import { SceneHeader } from '../../themed/SceneHeader'
+import { Tile } from '../../tiles/Tile'
 
 interface LocalState {
   selectedWallet: EdgeCurrencyWallet | null
