+++ src/components/scenes/Fio/FioDomainRegisterSelectWalletScene.tsx
@@ -5,23 +5,23 @@ import { ScrollView, View } from 'react-native'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 import { sprintf } from 'sprintf-js'
 
-import { FIO_STR } from '../../constants/WalletAndCurrencyConstants'
-import s from '../../locales/strings'
-import { getDomainRegInfo } from '../../modules/FioAddress/util'
-import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors'
-import { config } from '../../theme/appConfig'
-import { connect } from '../../types/reactRedux'
-import { RootState } from '../../types/reduxTypes'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { getWalletName } from '../../util/CurrencyWalletHelpers'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { ButtonsModal } from '../modals/ButtonsModal'
-import { WalletListModal, WalletListResult } from '../modals/WalletListModal'
-import { Airship, showError } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { EdgeText } from '../themed/EdgeText'
-import { MainButton } from '../themed/MainButton'
-import { Tile } from '../tiles/Tile'
+import { FIO_STR } from '../../../constants/WalletAndCurrencyConstants'
+import s from '../../../locales/strings'
+import { getDomainRegInfo } from '../../../modules/FioAddress/util'
+import { getDisplayDenomination, getExchangeDenomination } from '../../../selectors/DenominationSelectors'
+import { config } from '../../../theme/appConfig'
+import { connect } from '../../../types/reactRedux'
+import { RootState } from '../../../types/reduxTypes'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { getWalletName } from '../../../util/CurrencyWalletHelpers'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { ButtonsModal } from '../../modals/ButtonsModal'
+import { WalletListModal, WalletListResult } from '../../modals/WalletListModal'
+import { Airship, showError } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { EdgeText } from '../../themed/EdgeText'
+import { MainButton } from '../../themed/MainButton'
+import { Tile } from '../../tiles/Tile'
 
 interface StateProps {
   account: EdgeAccount
