+++ src/components/scenes/Fio/FioDomainSettingsScene.tsx
@@ -1,22 +1,22 @@
 import { EdgeCurrencyWallet } from 'edge-core-js'
 import * as React from 'react'
 
-import { refreshAllFioAddresses } from '../../actions/FioAddressActions'
-import { FIO_ADDRESS_DELIMITER } from '../../constants/WalletAndCurrencyConstants'
-import { formatDate } from '../../locales/intl'
-import s from '../../locales/strings'
-import { FioActionSubmit } from '../../modules/FioAddress/components/FioActionSubmit'
-import { getDomainSetVisibilityFee, getRenewalFee, getTransferFee, renewFioDomain, setDomainVisibility } from '../../modules/FioAddress/util'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { ButtonsModal } from '../modals/ButtonsModal'
-import { Airship, showError } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { ClickableText } from '../themed/ClickableText'
-import { EdgeText } from '../themed/EdgeText'
-import { MainButton } from '../themed/MainButton'
-import { Tile } from '../tiles/Tile'
+import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
+import { FIO_ADDRESS_DELIMITER } from '../../../constants/WalletAndCurrencyConstants'
+import { formatDate } from '../../../locales/intl'
+import s from '../../../locales/strings'
+import { FioActionSubmit } from '../../../modules/FioAddress/components/FioActionSubmit'
+import { getDomainSetVisibilityFee, getRenewalFee, getTransferFee, renewFioDomain, setDomainVisibility } from '../../../modules/FioAddress/util'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { ButtonsModal } from '../../modals/ButtonsModal'
+import { Airship, showError } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { ClickableText } from '../../themed/ClickableText'
+import { EdgeText } from '../../themed/EdgeText'
+import { MainButton } from '../../themed/MainButton'
+import { Tile } from '../../tiles/Tile'
 
 interface State {
   showRenew: boolean
