+++ src/components/scenes/Fio/FioNameConfirmScene.tsx
@@ -2,17 +2,17 @@ import { EdgeCurrencyConfig } from 'edge-core-js'
 import * as React from 'react'
 import { View } from 'react-native'
 
-import { FIO_ADDRESS_DELIMITER } from '../../constants/WalletAndCurrencyConstants'
-import s from '../../locales/strings'
-import { FioActionSubmit } from '../../modules/FioAddress/components/FioActionSubmit'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { ButtonsModal } from '../modals/ButtonsModal'
-import { Airship, showError } from '../services/AirshipInstance'
-import { cacheStyles, Theme } from '../services/ThemeContext'
-import { SceneHeader } from '../themed/SceneHeader'
-import { Tile } from '../tiles/Tile'
+import { FIO_ADDRESS_DELIMITER } from '../../../constants/WalletAndCurrencyConstants'
+import s from '../../../locales/strings'
+import { FioActionSubmit } from '../../../modules/FioAddress/components/FioActionSubmit'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { ButtonsModal } from '../../modals/ButtonsModal'
+import { Airship, showError } from '../../services/AirshipInstance'
+import { cacheStyles, Theme } from '../../services/ThemeContext'
+import { SceneHeader } from '../../themed/SceneHeader'
+import { Tile } from '../../tiles/Tile'
 
 interface StateProps {
   fioPlugin?: EdgeCurrencyConfig
