+++ src/components/scenes/Fio/FioRequestConfirmationScene.tsx
@@ -3,24 +3,24 @@ import { EdgeAccount, EdgeCurrencyConfig, EdgeCurrencyWallet } from 'edge-core-j
 import * as React from 'react'
 import { View } from 'react-native'
 
-import { formatNumber } from '../../locales/intl'
-import s from '../../locales/strings'
-import { addToFioAddressCache, checkPubAddress, convertEdgeToFIOCodes, getRemainingBundles } from '../../modules/FioAddress/util'
-import { CcWalletMap } from '../../reducers/FioReducer'
-import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors'
-import { getExchangeRate, getSelectedCurrencyWallet } from '../../selectors/WalletSelectors'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { emptyCurrencyInfo, GuiCurrencyInfo, GuiDenomination } from '../../types/types'
-import { DECIMAL_PRECISION, getDenomFromIsoCode } from '../../util/utils'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { AddressModal } from '../modals/AddressModal'
-import { ButtonsModal } from '../modals/ButtonsModal'
-import { TextInputModal } from '../modals/TextInputModal'
-import { Airship, showError, showToast } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { Slider } from '../themed/Slider'
-import { Tile } from '../tiles/Tile'
+import { formatNumber } from '../../../locales/intl'
+import s from '../../../locales/strings'
+import { addToFioAddressCache, checkPubAddress, convertEdgeToFIOCodes, getRemainingBundles } from '../../../modules/FioAddress/util'
+import { CcWalletMap } from '../../../reducers/FioReducer'
+import { getDisplayDenomination, getExchangeDenomination } from '../../../selectors/DenominationSelectors'
+import { getExchangeRate, getSelectedCurrencyWallet } from '../../../selectors/WalletSelectors'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { emptyCurrencyInfo, GuiCurrencyInfo, GuiDenomination } from '../../../types/types'
+import { DECIMAL_PRECISION, getDenomFromIsoCode } from '../../../util/utils'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { AddressModal } from '../../modals/AddressModal'
+import { ButtonsModal } from '../../modals/ButtonsModal'
+import { TextInputModal } from '../../modals/TextInputModal'
+import { Airship, showError, showToast } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { Slider } from '../../themed/Slider'
+import { Tile } from '../../tiles/Tile'
 
 interface StateProps {
   exchangeSecondaryToPrimaryRatio: string
