+++ src/components/scenes/Fio/FioRequestListScene.tsx
@@ -4,27 +4,27 @@ import * as React from 'react'
 import { ActivityIndicator, SectionList, View } from 'react-native'
 import { sprintf } from 'sprintf-js'
 
-import { refreshAllFioAddresses } from '../../actions/FioAddressActions'
-import { SPECIAL_CURRENCY_INFO } from '../../constants/WalletAndCurrencyConstants'
-import { formatDate, SHORT_DATE_FMT } from '../../locales/intl'
-import s from '../../locales/strings'
-import { addToFioAddressCache, cancelFioRequest, convertFIOToEdgeCodes, FIO_NO_BUNDLED_ERR_CODE } from '../../modules/FioAddress/util'
-import { getExchangeDenominationFromState } from '../../selectors/DenominationSelectors'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp } from '../../types/routerTypes'
-import { FioAddress, FioRequest } from '../../types/types'
-import { getTokenId } from '../../util/CurrencyInfoHelpers'
-import { tokenIdsToCurrencyCodes } from '../../util/utils'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { ButtonsModal } from '../modals/ButtonsModal'
-import { WalletListModal, WalletListResult } from '../modals/WalletListModal'
-import { FullScreenLoader } from '../progress-indicators/FullScreenLoader'
-import { Airship, showError, showToast } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { EdgeText } from '../themed/EdgeText'
-import { FioRequestRow } from '../themed/FioRequestRow'
-import { SceneHeader } from '../themed/SceneHeader'
-import { SectionHeader } from '../themed/TransactionListComponents'
+import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
+import { SPECIAL_CURRENCY_INFO } from '../../../constants/WalletAndCurrencyConstants'
+import { formatDate, SHORT_DATE_FMT } from '../../../locales/intl'
+import s from '../../../locales/strings'
+import { addToFioAddressCache, cancelFioRequest, convertFIOToEdgeCodes, FIO_NO_BUNDLED_ERR_CODE } from '../../../modules/FioAddress/util'
+import { getExchangeDenominationFromState } from '../../../selectors/DenominationSelectors'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp } from '../../../types/routerTypes'
+import { FioAddress, FioRequest } from '../../../types/types'
+import { getTokenId } from '../../../util/CurrencyInfoHelpers'
+import { tokenIdsToCurrencyCodes } from '../../../util/utils'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { ButtonsModal } from '../../modals/ButtonsModal'
+import { WalletListModal, WalletListResult } from '../../modals/WalletListModal'
+import { FullScreenLoader } from '../../progress-indicators/FullScreenLoader'
+import { Airship, showError, showToast } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { EdgeText } from '../../themed/EdgeText'
+import { FioRequestRow } from '../../themed/FioRequestRow'
+import { SceneHeader } from '../../themed/SceneHeader'
+import { SectionHeader } from '../../themed/TransactionListComponents'
 
 const SCROLL_THRESHOLD = 0.5
