+++ src/components/scenes/Fio/FioSentRequestDetailsScene.tsx
@@ -2,19 +2,19 @@ import { mul } from 'biggystring'
 import * as React from 'react'
 import { View } from 'react-native'
 
-import { FIAT_CODES_SYMBOLS } from '../../constants/WalletAndCurrencyConstants'
-import { formatDate, formatNumber, SHORT_DATE_FMT } from '../../locales/intl'
-import s from '../../locales/strings'
-import { isRejectedFioRequest, isSentFioRequest } from '../../modules/FioRequest/util'
-import { getSelectedCurrencyWallet } from '../../selectors/WalletSelectors'
-import { connect } from '../../types/reactRedux'
-import { RouteProp } from '../../types/routerTypes'
-import { GuiExchangeRates } from '../../types/types'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { EdgeText } from '../themed/EdgeText'
-import { SceneHeader } from '../themed/SceneHeader'
-import { Tile } from '../tiles/Tile'
+import { FIAT_CODES_SYMBOLS } from '../../../constants/WalletAndCurrencyConstants'
+import { formatDate, formatNumber, SHORT_DATE_FMT } from '../../../locales/intl'
+import s from '../../../locales/strings'
+import { isRejectedFioRequest, isSentFioRequest } from '../../../modules/FioRequest/util'
+import { getSelectedCurrencyWallet } from '../../../selectors/WalletSelectors'
+import { connect } from '../../../types/reactRedux'
+import { RouteProp } from '../../../types/routerTypes'
+import { GuiExchangeRates } from '../../../types/types'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { EdgeText } from '../../themed/EdgeText'
+import { SceneHeader } from '../../themed/SceneHeader'
+import { Tile } from '../../tiles/Tile'
 interface OwnProps {
   route: RouteProp<'fioSentRequestDetails'>
 }
