+++ src/components/scenes/Fio/FioStakingChangeScene.tsx
@@ -5,27 +5,27 @@ import { Image, View } from 'react-native'
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
 import { sprintf } from 'sprintf-js'
 
-import { refreshAllFioAddresses } from '../../actions/FioAddressActions'
-import fioLogo from '../../assets/images/fio/fio_logo.png'
-import { SPECIAL_CURRENCY_INFO, STAKING_BALANCES } from '../../constants/WalletAndCurrencyConstants'
-import { formatNumber, formatTimeDate, SHORT_DATE_FMT } from '../../locales/intl'
-import s from '../../locales/strings'
-import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors'
-import { convertCurrency } from '../../selectors/WalletSelectors'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { FioAddress } from '../../types/types'
-import { convertNativeToDenomination } from '../../util/utils'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { FlipInputModal, FlipInputModalResult } from '../modals/FlipInputModal'
-import { Airship, showToast } from '../services/AirshipInstance'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { EdgeText } from '../themed/EdgeText'
-import { ModalFooter, ModalMessage, ModalTitle } from '../themed/ModalParts'
-import { SceneHeader } from '../themed/SceneHeader'
-import { Slider } from '../themed/Slider'
-import { ThemedModal } from '../themed/ThemedModal'
-import { Tile } from '../tiles/Tile'
+import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
+import fioLogo from '../../../assets/images/fio/fio_logo.png'
+import { SPECIAL_CURRENCY_INFO, STAKING_BALANCES } from '../../../constants/WalletAndCurrencyConstants'
+import { formatNumber, formatTimeDate, SHORT_DATE_FMT } from '../../../locales/intl'
+import s from '../../../locales/strings'
+import { getDisplayDenomination, getExchangeDenomination } from '../../../selectors/DenominationSelectors'
+import { convertCurrency } from '../../../selectors/WalletSelectors'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { FioAddress } from '../../../types/types'
+import { convertNativeToDenomination } from '../../../util/utils'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { FlipInputModal, FlipInputModalResult } from '../../modals/FlipInputModal'
+import { Airship, showToast } from '../../services/AirshipInstance'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { EdgeText } from '../../themed/EdgeText'
+import { ModalFooter, ModalMessage, ModalTitle } from '../../themed/ModalParts'
+import { SceneHeader } from '../../themed/SceneHeader'
+import { Slider } from '../../themed/Slider'
+import { ThemedModal } from '../../themed/ThemedModal'
+import { Tile } from '../../tiles/Tile'
 
 interface OwnProps {
   route: RouteProp<'fioStakingChange'>
