+++ src/components/scenes/Fio/FioStakingOverviewScene.tsx
@@ -4,24 +4,24 @@ import * as React from 'react'
 import { Image, ScrollView, View } from 'react-native'
 import { sprintf } from 'sprintf-js'
 
-import { refreshAllFioAddresses } from '../../actions/FioAddressActions'
-import fioLogo from '../../assets/images/fio/fio_logo.png'
-import { getSymbolFromCurrency, STAKING_BALANCES } from '../../constants/WalletAndCurrencyConstants'
-import { useWatch } from '../../hooks/useWatch'
-import { formatNumber, formatTimeDate } from '../../locales/intl'
-import s from '../../locales/strings'
-import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors'
-import { convertCurrency } from '../../selectors/WalletSelectors'
-import { connect } from '../../types/reactRedux'
-import { NavigationProp, RouteProp } from '../../types/routerTypes'
-import { convertNativeToDenomination } from '../../util/utils'
-import { SceneWrapper } from '../common/SceneWrapper'
-import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
-import { ClickableText } from '../themed/ClickableText'
-import { EdgeText } from '../themed/EdgeText'
-import { MainButton } from '../themed/MainButton'
-import { SceneHeader } from '../themed/SceneHeader'
-import { Tile } from '../tiles/Tile'
+import { refreshAllFioAddresses } from '../../../actions/FioAddressActions'
+import fioLogo from '../../../assets/images/fio/fio_logo.png'
+import { getSymbolFromCurrency, STAKING_BALANCES } from '../../../constants/WalletAndCurrencyConstants'
+import { useWatch } from '../../../hooks/useWatch'
+import { formatNumber, formatTimeDate } from '../../../locales/intl'
+import s from '../../../locales/strings'
+import { getDisplayDenomination, getExchangeDenomination } from '../../../selectors/DenominationSelectors'
+import { convertCurrency } from '../../../selectors/WalletSelectors'
+import { connect } from '../../../types/reactRedux'
+import { NavigationProp, RouteProp } from '../../../types/routerTypes'
+import { convertNativeToDenomination } from '../../../util/utils'
+import { SceneWrapper } from '../../common/SceneWrapper'
+import { cacheStyles, Theme, ThemeProps, withTheme } from '../../services/ThemeContext'
+import { ClickableText } from '../../themed/ClickableText'
+import { EdgeText } from '../../themed/EdgeText'
+import { MainButton } from '../../themed/MainButton'
+import { SceneHeader } from '../../themed/SceneHeader'
+import { Tile } from '../../tiles/Tile'
 
 interface OwnProps {
   navigation: NavigationProp<'fioStakingOverview'>
